# Add fp16 support for faster inference