--- a/.gitignore
+++ b/.gitignore
@@ -1,0 +1,9 @@
+# ignore pycache
+__pycache__/
+
+# ignore generated videos
+*.mp4
+
+# ignore .egg-info from setup tools
+*.egg-info
+