## Usage:

Flag for half-precision / float16 inference:
```bash
python metadrive_simgen.py -p fp16
```
If not provided, the default setting will be float32.

## Comparison for inference speeds:
### **fp32:**
```
Running DDIM Sampling with 60 timesteps
Decoding image: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 60/60 [00:08<00:00,  7.23it/s]
Running DDIM Sampling with 100 timesteps
DDIM Sampler: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 100/100 [00:20<00:00,  4.96it/s]
```
60 timesteps:  **~7 it/s, 8 seconds**
100 timesteps:  **~5 it/s, 20 seconds**


### **fp16:** 
```
Running DDIM Sampling with 60 timesteps
Decoding image: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 60/60 [00:02<00:00, 20.56it/s]
Running DDIM Sampling with 100 timesteps
DDIM Sampler: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 100/100 [00:07<00:00, 13.69it/s]
```

60 timesteps:  **~20 it/s, 2 seconds**
100 timesteps:  **~14 it/s, 7 seconds**